refactor: Update group service to use int64 for group ID
Updated the GroupService to use int64 for the group ID in the GetGroupByID method. This change ensures consistency with the updated models and database schema, preventing potential data type mismatch errors.

+++ backend/internal/service/group_service.go
@@ -17,6 +17,6 @@ func (s *groupService) CreateGroup(group *models.Group) (*models.Group, error) {
 	return s.groupStore.CreateGroup(group)
 }
 
-func (s *groupService) GetGroupByID(groupID int) (*models.Group, error) {
+func (s *groupService) GetGroupByID(groupID int64) (*models.Group, error) {
 	return s.groupStore.GetGroupByID(groupID)
 }
